# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Design Bees Outbound Revenue Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
        - name: priorityTier
          in: query
          schema: { type: string }
        - name: icpType
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListResponse"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
  /accounts/import:
    post:
      operationId: importAccounts
      tags: [accounts]
      summary: Bulk import accounts from CSV row data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportAccountsRequest"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Account detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete an account
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ─── CONTACTS ───────────────────────────────────────────────────────────────
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: accountId
          in: query
          schema: { type: string }
        - name: outreachStatus
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactListResponse"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ─── ANALYSES ───────────────────────────────────────────────────────────────
  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List homepage analyses
      parameters:
        - name: accountId
          in: query
          schema: { type: string }
        - name: issueCode
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisListResponse"
    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Submit a page for analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalysisRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResponse"
  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get an analysis by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateAnalysis
      tags: [analyses]
      summary: Update an analysis result
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnalysisRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResponse"

  # ─── INSIGHTS ───────────────────────────────────────────────────────────────
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activities
      parameters:
        - name: accountId
          in: query
          schema: { type: string }
        - name: contactId
          in: query
          schema: { type: string }
        - name: activityType
          in: query
          schema: { type: string }
        - name: dueToday
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityListResponse"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Log an activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityResponse"

  # ─── OPPORTUNITIES ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
      required: [error]

    # Pagination wrapper
    PaginatedMeta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
      required: [total, page, pageSize]

    # ─── ACCOUNT ────────────────────────────────────────────────────────────
    Account:
      type: object
      properties:
        id: { type: string }
        companyName: { type: string }
        domain: { type: string, nullable: true }
        websiteUrl: { type: string, nullable: true }
        industry: { type: string, nullable: true }
        subIndustry: { type: string, nullable: true }
        employeeBand: { type: string, nullable: true }
        revenueBand: { type: string, nullable: true }
        geography: { type: string, nullable: true }
        icpType: { type: string, nullable: true }
        fitScore: { type: integer, nullable: true }
        priorityTier: { type: string, nullable: true }
        source: { type: string, nullable: true }
        sourceDetail: { type: string, nullable: true }
        sourceCaptureDate: { type: string, format: date-time, nullable: true }
        ownerId: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        contactCount: { type: integer, nullable: true }
      required: [id, companyName, ownerId, createdAt, updatedAt]
    AccountResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Account"
      required: [data]
    AccountDetailResponse:
      type: object
      properties:
        data:
          allOf:
            - $ref: "#/components/schemas/Account"
            - type: object
              properties:
                contacts:
                  type: array
                  items:
                    $ref: "#/components/schemas/Contact"
                analyses:
                  type: array
                  items:
                    $ref: "#/components/schemas/Analysis"
                recentActivities:
                  type: array
                  items:
                    $ref: "#/components/schemas/Activity"
      required: [data]
    AccountListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Account"
        meta:
          $ref: "#/components/schemas/PaginatedMeta"
      required: [data, meta]
    CreateAccountRequest:
      type: object
      properties:
        companyName: { type: string }
        domain: { type: string, nullable: true }
        websiteUrl: { type: string, nullable: true }
        industry: { type: string, nullable: true }
        subIndustry: { type: string, nullable: true }
        employeeBand: { type: string, nullable: true }
        revenueBand: { type: string, nullable: true }
        geography: { type: string, nullable: true }
        icpType: { type: string, nullable: true }
        fitScore: { type: integer, nullable: true }
        priorityTier: { type: string, nullable: true }
        source: { type: string, nullable: true }
        sourceDetail: { type: string, nullable: true }
      required: [companyName]
    UpdateAccountRequest:
      type: object
      properties:
        companyName: { type: string }
        domain: { type: string, nullable: true }
        websiteUrl: { type: string, nullable: true }
        industry: { type: string, nullable: true }
        subIndustry: { type: string, nullable: true }
        employeeBand: { type: string, nullable: true }
        revenueBand: { type: string, nullable: true }
        geography: { type: string, nullable: true }
        icpType: { type: string, nullable: true }
        fitScore: { type: integer, nullable: true }
        priorityTier: { type: string, nullable: true }
        source: { type: string, nullable: true }
        sourceDetail: { type: string, nullable: true }
    ImportAccountsRequest:
      type: object
      properties:
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [rows]
    ImportResult:
      type: object
      properties:
        imported: { type: integer }
        skipped: { type: integer }
        errors:
          type: array
          items: { type: string }
      required: [imported, skipped, errors]

    # ─── CONTACT ────────────────────────────────────────────────────────────
    Contact:
      type: object
      properties:
        id: { type: string }
        accountId: { type: string }
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        fullName: { type: string, nullable: true }
        jobTitle: { type: string, nullable: true }
        seniority: { type: string, nullable: true }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        linkedinUrl: { type: string, nullable: true }
        contactSource: { type: string, nullable: true }
        outreachStatus: { type: string }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        accountName: { type: string, nullable: true }
      required: [id, accountId, outreachStatus, createdAt, updatedAt]
    ContactResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Contact"
      required: [data]
    ContactListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        meta:
          $ref: "#/components/schemas/PaginatedMeta"
      required: [data, meta]
    CreateContactRequest:
      type: object
      properties:
        accountId: { type: string }
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        jobTitle: { type: string, nullable: true }
        seniority: { type: string, nullable: true }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        linkedinUrl: { type: string, nullable: true }
        contactSource: { type: string, nullable: true }
        notes: { type: string, nullable: true }
      required: [accountId]
    UpdateContactRequest:
      type: object
      properties:
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        jobTitle: { type: string, nullable: true }
        seniority: { type: string, nullable: true }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        linkedinUrl: { type: string, nullable: true }
        outreachStatus: { type: string }
        notes: { type: string, nullable: true }

    # ─── ANALYSIS ───────────────────────────────────────────────────────────
    Analysis:
      type: object
      properties:
        id: { type: string }
        accountId: { type: string }
        domain: { type: string }
        pageUrl: { type: string }
        pageType: { type: string }
        analyzedAt: { type: string, format: date-time }
        heroClarity: { type: integer, nullable: true }
        ctaClarity: { type: integer, nullable: true }
        ctaProminence: { type: integer, nullable: true }
        visualHierarchy: { type: integer, nullable: true }
        messageOrder: { type: integer, nullable: true }
        outcomeClarity: { type: integer, nullable: true }
        trustSignal: { type: integer, nullable: true }
        friction: { type: integer, nullable: true }
        mobileReadability: { type: integer, nullable: true }
        primaryIssueCode: { type: string, nullable: true }
        secondaryIssueCode: { type: string, nullable: true }
        tertiaryIssueCode: { type: string, nullable: true }
        issueSummaryShort: { type: string, nullable: true }
        issueSummaryDetailed: { type: string, nullable: true }
        recommendedPriorityFix: { type: string, nullable: true }
        confidenceScore: { type: integer, nullable: true }
        screenshotUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, accountId, domain, pageUrl, pageType, analyzedAt, createdAt]
    AnalysisResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Analysis"
      required: [data]
    AnalysisListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Analysis"
        total: { type: integer }
      required: [data, total]
    CreateAnalysisRequest:
      type: object
      properties:
        accountId: { type: string }
        domain: { type: string }
        pageUrl: { type: string }
        pageType: { type: string }
      required: [accountId, domain, pageUrl, pageType]
    UpdateAnalysisRequest:
      type: object
      properties:
        heroClarity: { type: integer, nullable: true }
        ctaClarity: { type: integer, nullable: true }
        ctaProminence: { type: integer, nullable: true }
        visualHierarchy: { type: integer, nullable: true }
        messageOrder: { type: integer, nullable: true }
        outcomeClarity: { type: integer, nullable: true }
        trustSignal: { type: integer, nullable: true }
        friction: { type: integer, nullable: true }
        mobileReadability: { type: integer, nullable: true }
        primaryIssueCode: { type: string, nullable: true }
        secondaryIssueCode: { type: string, nullable: true }
        tertiaryIssueCode: { type: string, nullable: true }
        issueSummaryShort: { type: string, nullable: true }
        issueSummaryDetailed: { type: string, nullable: true }
        recommendedPriorityFix: { type: string, nullable: true }
        confidenceScore: { type: integer, nullable: true }
        screenshotUrl: { type: string, nullable: true }

    # ─── INSIGHT ────────────────────────────────────────────────────────────
...
    Activity:
      type: object
      properties:
        id: { type: string }
        accountId: { type: string }
        contactId: { type: string }
        channel: { type: string }
        activityType: { type: string }
        subject: { type: string, nullable: true }
        body: { type: string, nullable: true }
        sentAt: { type: string, format: date-time }
        repliedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, accountId, contactId, channel, activityType, sentAt, createdAt]
    ActivityResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Activity"
      required: [data]
    ActivityListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        total: { type: integer }
      required: [data, total]
    CreateActivityRequest:
      type: object
      properties:
        accountId: { type: string }
        contactId: { type: string }
        channel: { type: string }
        activityType: { type: string }
        subject: { type: string, nullable: true }
        body: { type: string, nullable: true }
      required: [accountId, contactId, channel, activityType]

    # ─── OPPORTUNITY ────────────────────────────────────────────────────────
...