# lib/api-zod/src/generated/types/paginatedMeta.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Design Bees Outbound Revenue Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface PaginatedMeta {
  total: number;
  page: number;
  pageSize: number;
}