# lib/api-zod/src/generated/types/updateAnalysisRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Design Bees Outbound Revenue Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAnalysisRequest {
  heroClarity?: number | null;
  ctaClarity?: number | null;
  ctaProminence?: number | null;
  visualHierarchy?: number | null;
  messageOrder?: number | null;
  outcomeClarity?: number | null;
  trustSignal?: number | null;
  friction?: number | null;
  mobileReadability?: number | null;
  primaryIssueCode?: string | null;
  secondaryIssueCode?: string | null;
  tertiaryIssueCode?: string | null;
  issueSummaryShort?: string | null;
  issueSummaryDetailed?: string | null;
  recommendedPriorityFix?: string | null;
  confidenceScore?: number | null;
  screenshotUrl?: string | null;
}